substantial rewrite, add support for svg:polygon
With this patch we do a full traversal of the svg document, rather than
an xpath lookup to just pull out one kind of element.  It places the
program in a better position to be able to add support for more element
types.

We will need to re-add support for svg:path point sequences and scaling.