eclipse-path: program to produce vla model of an eclipse path
This program expects eclipse data in a particular format that was copied
from a nasa web page.  I don't think it is a standard format, so that
makes this program fairly fragile.  I'm committing the data until such
time as we can make this program work with more easily obtainable data
in some standard format.