eclipse-path: latlon->cartesian3 result in more conventional coordinate system
The VLA-LEFT coordinate system is X right, Y up, Z forward.

Digistar's WCS is more conventional: X right, Y forward, Z up.

The Dspath format uses WCS, so this patch changes the definition of
latlon->cartesian3 to give its result in WCS.  Output-vla is modified to
swap the y and z values.

This patch also makes a trivial change to the order of the math in
latlon->cartesian3 for readability.