eclipse-path: get rid of jd global
This is just a step toward solving this problem.  The problem is that
the catalog doesn't contain any dates, only times, so we somehow need to
convey that information to the catalog parser.  This patch just shifts
things around a bit to put the code in a better form for solving this
problem in the future.